+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -450,6 +450,7 @@ sub build_isf_public_interface_contract {
             't/1281-isf-enum-member-library-use-params.t',
             't/1282-isf-enum-member-drive-expression-values.t',
             't/1283-isf-aggregate-rule-values.t',
+            't/1284-isf-aggregate-rule-expression-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',